requirements.txt
mediapipe
opencv-python
uvicorn
fastapi
numpy
pyspellchecker
keras
